Notif Error + Success Import Prodi Prestasi

+++ app/Http/Controllers/ProdiController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Imports\ProdiImport;
 use App\Models\Criteria;
 use App\Models\Prodi;
+use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 
@@ -16,53 +17,54 @@ class ProdiController extends Controller
 
     public function import (Request $request) 
     {
-        $array = (new ProdiImport())->toArray($request->file('excel'));
-
-        $namedkey = array();
-        for ($i=0; $i < $request->input('banyakCollumn'); $i++) {
-            $namedkey[$i]=strtolower($request->input('collumn-'.strval($i)));
+        if ($request->file('excel') == null ||
+            $request->input('periode') == '' ||
+            $request->input('banyakCollumn') == 0) {
+                Session::flash('error','Pastikan anda telah mengisi semua input');
+                return redirect()->back();
         }
 
-        // $namedkey = array(
-        //     strtolower($request->input('col_id_prodi')), 
-        //     strtolower($request->input('col_jurusan')), 
-        //     strtolower($request->input('col_id_politeknik')), 
-        //     strtolower($request->input('col_politeknik')), 
-        //     strtolower($request->input('col_id_kelompok_bidang')), 
-        //     strtolower($request->input('col_kelompok_bidang')), 
-        //     strtolower($request->input('col_Quota')), 
-        //     strtolower($request->input('col_tertampung'))
-        // );
+        try {
+            $array = (new ProdiImport())->toArray($request->file('excel'));
 
-        $periode = $request->input('periode');
-
-        $criteria = array(
-            'tahun' => $periode,
-            'criteria' => $namedkey,
-            'table' => 'prodi',
-            'kode_criteria' => strval($periode).'_prodi',
-        );
-        if (Criteria::where('kode_criteria',strval($periode).'_prodi')->first()) {
-            Criteria::where('kode_criteria',strval($periode).'_prodi')->update($criteria);
-        } else {
-            Criteria::insert($criteria);
+            $namedkey = array();
+            for ($i=0; $i < $request->input('banyakCollumn'); $i++) {
+                $namedkey[$i]=strtolower($request->input('collumn-'.strval($i)));
+            }
+    
+            $periode = $request->input('periode');
+    
+            $criteria = array(
+                'tahun' => $periode,
+                'criteria' => $namedkey,
+                'table' => 'prodi',
+                'kode_criteria' => strval($periode).'_prodi',
+            );
+    
+            for ($i=0; $i < count($array[0]); $i++) {
+                // $fil = array();
+                for ($ab=0; $ab < count($namedkey); $ab++) { 
+                    $fil[$namedkey[$ab]] = trim($array[0][$i][$namedkey[$ab]]);
+                };
+                $fil['periode'] = $periode;
+                $filtered[] = $fil;
+            }
+    
+            if (Criteria::where('kode_criteria',strval($periode).'_prodi')->first()) {
+                Criteria::where('kode_criteria',strval($periode).'_prodi')->update($criteria);
+            } else {
+                Criteria::insert($criteria);
+            }
+    
+            Prodi::truncate();
+            Prodi::insert($filtered);
+            
+            Session::flash('sukses','Data Calon Mahasiswa Berhasil diimport');
+            return redirect()->back();
+        }catch (Exception $error) {
+            Session::flash('error', $error);
+            return redirect()->back();
         }
-
-
-        for ($i=0; $i < count($array[0]); $i++) {
-            // $fil = array();
-            for ($ab=0; $ab < count($namedkey); $ab++) { 
-                $fil[$namedkey[$ab]] = trim($array[0][$i][$namedkey[$ab]]);
-            };
-            $fil['periode'] = $periode;
-            $filtered[] = $fil;
-        }
-
-        Prodi::truncate();
-        Prodi::insert($filtered);
-        
-        Session::flash('sukses','Data Berhasil ditambahkan');
-        return redirect()->back();
     }
 
     public function render()
